+++ packages/data/src/entities/placeline.ts
@@ -4,7 +4,7 @@ import {
   IActivitySegment,
   IEventMark,
   IPlacelineSegment,
-  IProcSegment
+  IProcSegment, ISegmentType
 } from "../interfaces";
 import { htEvent } from "./event";
 import { NameCase, GetMinute } from "ht-utility";
@@ -12,7 +12,7 @@ import _ from "underscore";
 import { CommonFunctions } from "../common";
 
 export class HtPlaceline {
-  constructor(public userData: IUserData) {}
+  constructor(public userData?: IUserData) {}
 
   getPlacelineSegments() {
     let userData = this.userData;
@@ -134,6 +134,28 @@ export class HtPlaceline {
       segmentId: lastSeg.id,
       ...this.getActivityStyleClass(lastSeg)
     };
+  }
+
+  isLive(placeline: IUserData) {
+    let old = placeline.display.seconds_elapsed_since_last_heartbeat;
+    let status = placeline.display.status_text;
+    return status !== 'Logged off' && old < 15 * 60;
+  }
+
+  getSegmentTypes(userSegments: ISegment[]) {
+    return _.reduce(
+      userSegments,
+      (segmentType: ISegmentType, segment: ISegment) => {
+        if (segment.type == "stop") {
+          if (segment.location && segment.location.geojson)
+            segmentType.stopSegment.push(segment);
+        } else {
+          if (segment.encoded_polyline) segmentType.tripSegment.push(segment);
+        }
+        return segmentType;
+      },
+      { tripSegment: [], stopSegment: [] }
+    );
   }
 
   //helpers
@@ -277,4 +299,4 @@ export class HtPlaceline {
   }
 }
 
-export const htPlaceline = (userData: IUserData) => new HtPlaceline(userData);
+export const htPlaceline = (userData?: IUserData) => new HtPlaceline(userData);